Unregister the service worker when client_side_replay is off (#968)
If you have client_side_replay on and later turn it off,
prior visitors will still have it registered in their browser until it expires unless they manually remove it through devtools or by clearing site data. While we don't expect to be regularly flipping it on and off in production, it's important to be able to roll back to the previous configuration if an issue is discovered.